# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchDeleteRequest": {
        "description": "批量删除请求。",
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "BatchDeleteRequest",
        "type": "object"
      },
      "BatchDeleteResponse": {
        "description": "批量删除响应。",
        "properties": {
          "deleted_count": {
            "title": "Deleted Count",
            "type": "integer"
          }
        },
        "required": [
          "deleted_count"
        ],
        "title": "BatchDeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubmissionCreateResponse": {
        "description": "上传成功后返回最小信息。",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "SubmissionCreateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SURF-2026-0031 后端 API",
    "title": "AI 作业批改系统",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/submissions": {
      "delete": {
        "description": "批量删除提交记录。\n\n- 仅终态记录允许删除;存在处理中记录时整批原子拒绝\n- 先提交数据库删除,成功后再清理磁盘 PDF(学生作业 + 题目)\n- 关联的 conversations 由 DB CASCADE 自动级联删除\n- 不存在的 ID 安全跳过,不影响其他记录的删除\n- 文件删除失败不阻断流程(如文件已被清理),仅记录日志",
        "operationId": "batch_delete_submissions_api_submissions_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Delete Submissions",
        "tags": [
          "submissions"
        ]
      },
      "get": {
        "description": "分页列出提交(按上传时间降序)。\n\n``include_count=false`` 时跳过 ``COUNT(*)`` 查询,``total`` 返回 ``0``。\n用于前端处理中轮询场景:轮询只需刷新 items,总数由独立的\n``GET /submissions/count`` 提供,避免每 15s 一次 count 往返。",
        "operationId": "list_submissions_api_submissions_get",
        "parameters": [
          {
            "description": "跳过的记录数",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "跳过的记录数",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "每页记录数",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "每页记录数",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "是否计算 total。轮询场景传 false 跳过 count 查询",
            "in": "query",
            "name": "include_count",
            "required": false,
            "schema": {
              "default": true,
              "description": "是否计算 total。轮询场景传 false 跳过 count 查询",
              "title": "Include Count",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSubmissions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Submissions",
        "tags": [
          "submissions"
        ]
      },
      "post": {
        "description": "上传学生报告 PDF 及可选的多语言代码文件。\n\n收敛为 MCP-only 后固定使用外部编程助手评分模式：上传后只做 OCR，\n进入 ``awaiting_mcp``，由 MCP 客户端评分。",
        "operationId": "create_submission_api_submissions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_submission_api_submissions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Submission",
        "tags": [
          "submissions"
        ]
      }
    },
...
    "/api/submissions/export.xlsx": {
      "get": {
        "description": "Export teacher-reviewed results for one question as an Excel workbook.",
        "operationId": "export_submission_results_api_submissions_export_xlsx_get",
        "parameters": [
          {
            "description": "题目 ID",
            "in": "query",
            "name": "question_id",
            "required": true,
            "schema": {
              "description": "题目 ID",
              "minLength": 1,
              "title": "Question Id",
              "type": "string"
            }
          },
          {
            "description": "本次导出的及格得分率百分比",
            "in": "query",
            "name": "pass_threshold",
            "required": true,
            "schema": {
              "description": "本次导出的及格得分率百分比",
              "exclusiveMinimum": 0,
              "maximum": 100,
              "title": "Pass Threshold",
              "type": "number"
            }
          },
          {
            "description": "工作簿标题语言",
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "zh-CN",
              "description": "工作簿标题语言",
              "enum": [
                "zh-CN",
                "en-US"
              ],
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Submission Results",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/submissions/{submission_id}": {
      "get": {
        "description": "获取单个提交详情。",
        "operationId": "get_submission_api_submissions__submission_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Submission",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/submissions/{submission_id}/events": {
      "get": {
        "description": "SSE 端点:实时推送 submission 状态变更事件(P1)。\n\n- 基于 PostgreSQL LISTEN/NOTIFY,worker 在 ``_update_status`` 内 NOTIFY\n- 启动时先推一次当前状态,避免客户端错过终态事件\n- 每 15s 注释行 keepalive,防止中间代理断开空闲连接\n- 前端 ``useSubmissionEvents`` hook 消费此端点,收到事件后 invalidate\n  status query;30s 兜底轮询在 SSE 断开时仍能恢复",
        "operationId": "submission_events_api_submissions__submission_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submission Events",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/submissions/{submission_id}/finalize": {
      "post": {
        "description": "教师确认最终评分。\n\n- 校验状态为 ready_for_review(已 reviewed 返回 409)\n- 校验分数一致性(由 FinalizeRequest 的 model_validator 处理)\n- 写入 score/max_score/feedback/details/reviewed_by/reviewed_at\n- 状态置为 reviewed",
        "operationId": "finalize_submission_api_submissions__submission_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Submission",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/submissions/{submission_id}/pdf": {
      "get": {
        "description": "安全返回学生作业或作业题目的 PDF 文件流。\n\n- 仅终态记录(ready_for_review/reviewed/failed)可访问\n- 不暴露服务器文件路径,仅通过 DB 读取后本地读取\n- 使用浏览器原生 PDF 预览(iframe/object)",
        "operationId": "get_submission_pdf_api_submissions__submission_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          },
          {
            "description": "PDF 类型: submission 或 question",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "submission",
              "description": "PDF 类型: submission 或 question",
              "title": "Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Submission Pdf",
        "tags": [
          "submissions"
        ]
      }
    },
...
    "/api/submissions/{submission_id}/status": {
      "get": {
        "description": "仅返回当前提交的处理状态与终态字段。\n\n用于前端兜底轮询场景(主路径走 SSE 推送):每 30s 拉一次轻量接口,\n进入终态后再拉完整 ``GET /submissions/{id}``。payload 体积通常\n小于 200B,远小于完整 ``SubmissionDetail`` 含 ``ocr_text`` 的几十 KB 至 MB。",
        "operationId": "get_submission_status_api_submissions__submission_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Submission Status",
        "tags": [
          "submissions"
        ]
      }
    }
  }
}